Add Swagger docs to api layer
Added swagger docs to the api layer. Run the swagger with make file
command. You have to serve the yaml file yourself atm. Its not hooked up
with any UI operation.

Users are able to view the swagger however interactions might be limited
since they are not served to a local platform.

closes #96