Add storybook for inventories and update notes
- add missing storybook for inventories component
- add dialog and drawer component storybook
- remove react calandar
- update volunteering activities to pass +ve numbeR
- remove immer for notes section
- removes immer for the notes section and uses the object.assign process
to make the notes function as they should in the profile page. the
issue was that after sucessfully editing a note, produce updated the
original object and hence next time we load the same button anywhere
else in the page, we are getting stale context. that is because the
immer updated original object from constants file but we have not
removed the modal from the DOM. tried many ways, was not able to remove
the modal from the DOM. instead of many states solution or ref solution,
all we do is everytime we press cancel or submit, the entire state is
wiped clean and everything else must be reloaded.

- display event creator and created time
- add storybook for inventories
- add helper text to outline volunteering hours are rounded up
closes #112 #141