Update inventory list and storybook cleanup
- update inventory list to display id of bulk downloaded items as well
- update storybook component to display props
- update the application validator to support user workflow

if the token is expired, we get error from the backend service. this
change will redirect the usr to the landing page if they pressed
refresh on the page. unless they press refresh they still can traverse
around the page. this is intended. if we can apply the patch fix to
redirect user to the landing page that would be great but the
application is not built like that. the landing page is based on the
state of the userID. this state although removed from the local storage,
react doesn't re-render the application validator and therefore the user
has to refresh. as far as i can tell, the api can throw an error, and if
the error is caught we can redirect the user to the landing page, but
that has to be done for every single api or a middleware must be
introduced. this is the patch fix we can go after but for now this
should suffice.

- minor tech debt changes
- adding better unit test to support the workflow of associated
  inventories

adding test to create an event and test if the incorrect user can
attempt to reach the shared inventories.

- removed validation of uuid check in db rather put it outside in the
  handler

i have noticed that we are doing this in multiple places and cleanup
should also be associated there. TODO :- https://github.com/earmuff-jam/mashed/issues/151
creatd a ticket to handle this issue.

- added new constants for share items component to support headers
  instead of writing them out one at a time.
- corrected prop types for ShareItemsComponent as well.