Add more fields on a selected item
- add ability to return an item

Add extra fields for few required fields

- add min and max for weight and height

Add more fields in testdata.sql to associate with a selected event

Add minor backend changes to support more inv fields

- added sample data in testdata.sql
- update existing sql to support more fields
- added ability to filter unwanted columns in the inventory list
- added similar design as shared item component for consistency and bring light on a seperate edit mode for select item.
- added easy edit for toggle selection on returnable field, others stay same. all new fields are editable.
- updated constants to support extra fields for header purposes.
- added more fields to inventory struct
- added the fields in each of the response.

we should probably refactor the sql since we are using the same thing a couple of times. it would be nice to update it only at one location. some of the api fields do not return the added extra fields and finding that and adding that is getting difficult. we should consider to consolidate if possible

flush headers clearly