160 tech debt cleanup (#161)
- Effort to migrate toward muiv5 emotion
A lot of components have been changed, but we are starting to roll out muiv5 for all components.
We want to also take this time to make each page better since there is not a lot of those either.
This PR should start with the Authentication landing page.

- attempted to broadcast the user count in for all users
We might have to put this feature behind a feature flag at the moment since I'm not quite sure why
the app is not working as expected. We need some time to dig into this at the moment.

- add linter to cleanup a lot of files
Adding linter to cleanup a lot of files. these changes are small but noteworthy as it would make the
app much better to develop against and use.

- add new folder paradigm of features
creating a lot of components and containers gave us a lot of difficulty in finding the code for each
respective component. we want to reduce this overhead. combined all items into features folder. each page
can get its own feature component. we also want to seperate out the /edit components but the routes are not
built like that just yet. we might have to facility more changes to support that.

- added detailed dev setup guide for users to run the application
added much more detailed support to build the application for our devs. this gives the leverage to quickly
setup an env and build on it.

There is just too much to document changes, but a lot of them are just minor code changes and tech debt updates.
Muiv5 migration is continued in tickets further down the line.

Closes #160