Tech debt cleanup
- continue migration effort to muiv5
- added appbar for minor action
- added qr code generator pacakge and color input pacakges to support
color and qr code generation of selected inventory item
- added collection on bottom on page to quickly navigate to various
fields of the app like creating new categories, maintenance plans etc
- add extra navigation rows
- add extra navigation rows for categories, plans
- add navigation for reports
- continue with the muiv5 migration
- revamp landing page and profile page to muiv5
- revamps landing page to muiv5
- continue on the effort to migrate towards muiv5
- revamps profile page to muiv5
- revamps notes page to muiv5
- migrate more files towards muiv5 components.
- remove eslint warning and suppress for a few files.
- fix all other warning and errors for files.
- add fix for backend test

With the profile changes, we noticed a couple of files that do not need
to be included and hence removing them. There are some cleanup effort to
do in the api layer as well.

Added profile details page where users can edit on the fly. No need to
press edit button. This cleans up a lot of discrepancies in the profile
table as well. We removed roles and online status as it not much of use
anyways. Somehow the unit test is not working as expected, but i'm not able
to repro it. going to let the pipeline tell me if its failed.

New navigation allows to select personal notes from the side easily. Updated
code for new notes but uses existing older layout.