Seperate inventory and add dark mode for app
- add application support for dark mode
- seperate inventory to its own saga and slice
- support for dark mode throughout the app. this can be set currently from
the profile section we should move it to the main app bar so users have easy access.
- unit test fix for profile db
- add fixes for lint issue
- add fix for editing an inventory in ui layer. updated slice so that both inventories
and inventory in the redux store can get populated as we need the values overall.
- support for editing inventory in db layer. currently all api calls are required
to put everything back. we still need to support updating one item so keeping some of
the existing api for later use.

Todo :
1. add support for return date time. this doesn't exist in the db. we need to support this.
2. add unit test to cover all grounds.