Add location picker to categories component (#174)
- adds location picker to categories
- add min, max count of items and status to categories component

add support for xs form factor

Refactor card layout for categories component

- add status selection options
- update card component for categories section
- refactor emptyComponent name
-

Create status table for item statuses in app

- add status to categories component
- add status to ntoes component
- add status to test_data.sql
- update test for backend features

Add status list api with tests

- add ability to retrieve status list api for options dropdown in
categories and notes component.
- add unit test to support get status list
- update get notes api to support for status list
- seperate retrieve selected status for notes. currently, the user selects
a status and we send the status title back. the status title cannot be
created at this time but in later date that might change. if such, we want
to ensure that the title is unique. it does not make sense to have two status
of the same name.
- added test coverage to support for this change in notes db.
- updated constants file to contain status label vs status display.
- added support to create, update and remove categories with max item and min item
- added support for colors for categories and added backend unit test for them
- refactored test_data.sql for categories and notes
- refactored unit test for notes db