203 - Add ability to upload images (#226)
Update docker compose file

- update docker compose file to support housing of local_images folder
in the application. this will allow us to keep uploaded images seperate that
the application itself.

Add filePath directory in main api.go layer

- the go service will serve up the file server api.
this will support loading and uploading images to and from the
local_uploads folder.

Added support to read images from volume

- add go api to retrieve files from local_uploads folder
- add local_uploads folder to gitignore so that we don't mistakely upload images
- added two api endpoints to return images in the filesystem. this part will require more cleanup.
currently files are stored within local_uploads/userID. if the user id changes we also need to
be able to update that, and we also need to be able to know / limit the size of the uploads. Furthermore,
when the container is closed, we need to make sure images are removed so the space is saved and this also needs
to work in other docker env files. currently support is built for local dev setup. later on, we will need the same support
in prod env.

Add permission script and add new apis

- added new permission script to allow read / write access to
local_uploads folder. since docker uses root as the main user, we need to
add the proper permissions so it can upload data into the local_uploads folder.

Update existing create api for file system

- altered api to create the new file inside the local_uploads folder.

Build frontend for image picker

- added ability to upload files and view them from the backend
- added saga and redux actions for image
- added support with formData for images.
- update api layer to support association handler for images
- update integration for inventories

Create sql file to support image url

- added new sql script to support images upload
- added image and display in inventories

Tech debt

- tech debt cleanup
- add redirect to home after error has occured
- new swagger changes

Address displaying images for inventories

- Updated api call in GridComponent to support only if there is an associated item url.
- Added takeEvery for redux action so that we don't leave behind the selected images for certain
components.
- Add 2mb limit to image upload
- Add check for filetype as well