233 Added username during signup (#237)
* 233 Add user validation during sign up

- added ability to create username during signup process
- added service layer to validate user email if they exist
- added unit test to support validation of user email
- addded model and perform tech debt cleanup
- added user model for user.go
- updated redux actions and saga for validation of user email address
- updated ui component to structure the check mark vs cross mark for valid user in the ui

closes #233

* fix unit tests

* Add t.skip to some of the tests

Skipping some of these tests since they conflict with the overall user
creation process. since the workflow is from handler to db, and db requires
certain users to be present which are only configured by the .env file.
these unit test pass on local but fail in the remote. that is because there is no
.env in the remote. we need to split these tests out into db and handler so that
there is not a dependency on the setup and creation of psql db unlike how we have it now.
skipping these tests until we find a resolution for it.