178 - Updated location picker component
Added ability to edit selected location

- added ability to change location for the selected location picker.
- added a new ref and a handler that listens to the click and returns the lat, lon
of the selected location. this allowed us to change the location from the map.

i think we should kind of enlarge the map a little bit since it is now difficult to traverse
around the map and if we add features to this map later on, we run the risk of having too much stuff
on a simple box.

Added location picker for Categories

- added location picker for categories component
- added new sql to add new column for categories component

Support multiple rows

- added support to display multiple categories and maintenance plans.
the current limit is at 100, if the user creates / has more items that 100,
we should allow the users to fetch more data.