251 - Update support for collaborators
Added fix to support adding / removing collaborators

- When users assign items into a select category or maintenance plan,
the underlying collaborators for the selected category / maintenace plan
will now also be applied to the selected item.
- When users assign new collaborators to already setup category or maintenace plans,
the collaborators will be assigned via trigger action of psql. This allows users to be
updated when the category is fully built out and can support changes throughout.

Closes #251