Update collaborators for category and plans
Attempted to fix the issue where for a second user after removing ownself
from collaborators, we realized that the page lands at the landing page, but
still shows up the previous category. we need to be able to fix this.

Also noted that using updatedBy during this update we pass in the current logged
in userID. and the creatorID is used to retrieve selected status, since it is mapped
to when it was initally created.