251 - Update support for collaborators (#252)
* 251 - Update support for collaborators

Added fix to support adding / removing collaborators

- When users assign items into a select category or maintenance plan,
the underlying collaborators for the selected category / maintenace plan
will now also be applied to the selected item.
- When users assign new collaborators to already setup category or maintenace plans,
the collaborators will be assigned via trigger action of psql. This allows users to be
updated when the category is fully built out and can support changes throughout.

Closes #251

* addressed icon issue during register

* Address SharableGroups issue for collaborators

 - Addressed issue where collaborators remove themslves from the selected category / maintenance plan. If the user removes themselves from the collaborators
 then the application will attempt to redirect the user to the main landing page since sometimes users might not have list of categories. Similar option is
 selected for users who remove themselves from collaborators status for maintenance plans.

* Update collaborators for category and plans

Attempted to fix the issue where for a second user after removing ownself
from collaborators, we realized that the page lands at the landing page, but
still shows up the previous category. we need to be able to fix this.

Also noted that using updatedBy during this update we pass in the current logged
in userID. and the creatorID is used to retrieve selected status, since it is mapped
to when it was initally created.