229 - Cleanup and Update Component Styling (#258)
* 229 - Update layout and update components

- Remove storybook components
Removing storybook components since they are not relevant
anymore. We have made numerous changes and none of them have
been a part of the new style changes. We need to update docs
and later on we can think about adding storybook files.

- Updated MenuActionBar component for better styling
Updating MenuActionBar component so that we can have selection color displayed
when we select a navigation menu item. Removed Settings and AppearanceSettings files
since they are all combined into Profile page now. We also added the ability to view
primary selected item with the theme colors. We used the react router path params to
determine the selected component.

- Update maxHeight for GridComponent, Categories component and Maintenance Plan component.
When we have a lot of data, the component was overflowing the screen and was getting difficult to
view the selections. We added maxHeight property of 65vh so that users are able to view the list
while still being able to see the header components.

- Tech debt cleanup
We performed some tech debt and css cleanup. Since we are adjusting for new layout, there
is a lot of cleanup to be done on the css aspect of the application. Of course, layouts and
design is always changing, but we want to keep things uniform and bring meaning to the app.
Also profile details was not supporting 500 characters limit for about me, so we updated that
as well.

Closes #229

* Add cleanup