Added support to update item details component
- Added ability to return error key if the requested image does not exist.

We did this to ensure that if the requested image was not present we would display
a placeholder. We did this by returning an error from the db, and using the redux to throw
an error. If the error happens, we automatically can display the placeholder image without much
difficulty.

- Added support to wrap the Graph content for both the maintenance item and category item details.

We wanted to have a empty placeholder when the graph was not displaying. Having an empty graph that
served no purpose and it just was taking real estate in the page. Instead we added the placeholder component
so that if there is no graph to display it would rather just display the Sorry no matching records found
empty placeholder component.

- Created storybook for the graph content for the maintenace item.

Since we are working on the maintenace item changes, we decided that it would be beneficial to have a
storybook dedicated to the contents of the graph and have a dedicated wrapper for the graph container itself
although it was only using a Paper component. Doing this, we are able to view the component easily in storybook
and it also looks much cleaner all in all.