Updated subtitle text for empty caption
Added relevant message during rendering of empty component
to ensure that the customers are aware of what they are seeing or in this
case what they are not seeing. For eg, if the graph is not present, we render
a EmptyComponent and a caption that states "Associate Assets" that reminds
users that they need to associate assets to view the graph.