Tech Debt Cleanup
- Added changes to props for TableComponent and maintenance items.

This was long overdue, a lot of props makes little sense to the developer, and
in order to shine better light on the props, we tried to refactor the maintenace item
section to properly use a good naming convention. This should also be carried out in
maintenance plan list, categories and its categories item because they all share similar
strategies.