290 - Add storybook for maintenance plan list
Added storybook for PlanHeader

Seperated the button component and the header component so we could
test properly with storybook. Created storybook for both the files and
addressed the import path for them. Instead of using relative import we
just want to use project specific imports like with @. This has been
an ongoing effort and more than likely will continue.

Added storybook for maintenance plan list

Added storybook for various scenarios within maintenance plan list

Created Add Maintenance Plan List

We seperated the add maintenance plan form for easier workflow.
Created seperate components for form fields and seperate for ColorPicker
components.

Tech debt and lint fixes

Had a couple of tech debt and lint issues to fix. Also renamed files for
maintenance plans. I think we attempted to remove the graph header, but
realized that when displaying empty maintenance plans, there is no indication
of what that section is. Hence we need the title.

Update maintenance plan to use status options

Removed association for maintenance plan with the status options. This
path was choosen because there was already enough statuses to keep maintenance
plans with. We didn't support adding new status for either the category, notes
or even for maintenance plan, so in order to remove overhead, we removed the
status options. In doing this, we removed all associations from status options
in the UI. Note, the backend uses the status options within other sections such as
Notes, Categories and cleanup will commence there when the time comes.

Updated Instances.js file

Instead of erroring out in all scenarios we split these between production and dev
environments. So now, if an error occurs in the production env, then the user will be
logged out for safety reasons and if we are in developer mode ( seperate docker container )
then we will be able to continue operation. This better helps to triage bugs and fix
defects easily and does not require the developer to continuously log in.

Closes #290