Added checked status to checkbox
In the CustomTableComponent file, we had a problem with
the checkbox and its rendering cycle. The issue was that after we selected
couple of fields, if we would select the main checkbox on the column, it would
act like it was unselecting all, and selecting all but the view showed otherwise.
In order to combat this wonkyness, we decided to add a checked status in the main
column checkbox which would let the user select a couple of rows and by doing this, it
would highlight checkbox of the column. When you select the column checkbox, it would now
unselect as expected. This keeps the behavior consistent with the expected outcome.