289 - Added storybook for maintenance item (#297)
* 289 - Added storybook for maintenance item

- Updated maintenance item component to use storybook.
- Created storybook for all other respective components.

- Updated styles for BarChart.jsx file.

Added background color to match the theme for item association graph
and removed border so that it looks flush with the theme.

Closes #289

* Added support to update item details component

- Added ability to return error key if the requested image does not exist.

We did this to ensure that if the requested image was not present we would display
a placeholder. We did this by returning an error from the db, and using the redux to throw
an error. If the error happens, we automatically can display the placeholder image without much
difficulty.

- Added support to wrap the Graph content for both the maintenance item and category item details.

We wanted to have a empty placeholder when the graph was not displaying. Having an empty graph that
served no purpose and it just was taking real estate in the page. Instead we added the placeholder component
so that if there is no graph to display it would rather just display the Sorry no matching records found
empty placeholder component.

- Created storybook for the graph content for the maintenace item.

Since we are working on the maintenace item changes, we decided that it would be beneficial to have a
storybook dedicated to the contents of the graph and have a dedicated wrapper for the graph container itself
although it was only using a Paper component. Doing this, we are able to view the component easily in storybook
and it also looks much cleaner all in all.

* Updated subtitle text for empty caption

Added relevant message during rendering of empty component
to ensure that the customers are aware of what they are seeing or in this
case what they are not seeing. For eg, if the graph is not present, we render
a EmptyComponent and a caption that states "Associate Assets" that reminds
users that they need to associate assets to view the graph.

* Add cleanup when selected image is not present

Fixed the issue where selecting between two different maintenance plans,
the previous image would persist and hence display the wrong image. Updated
the redux action to remove image if there is no image in the container.

Tech debt changes

Fixed props to TableComponent as we were missing a single prop.

* Tech Debt Cleanup

- Added changes to props for TableComponent and maintenance items.

This was long overdue, a lot of props makes little sense to the developer, and
in order to shine better light on the props, we tried to refactor the maintenace item
section to properly use a good naming convention. This should also be carried out in
maintenance plan list, categories and its categories item because they all share similar
strategies.

* add gap between row header and content in maintenance item components

* Add support to add inventory list into a maintenance plan item list

* Added checked status to checkbox

In the CustomTableComponent file, we had a problem with
the checkbox and its rendering cycle. The issue was that after we selected
couple of fields, if we would select the main checkbox on the column, it would
act like it was unselecting all, and selecting all but the view showed otherwise.
In order to combat this wonkyness, we decided to add a checked status in the main
column checkbox which would let the user select a couple of rows and by doing this, it
would highlight checkbox of the column. When you select the column checkbox, it would now
unselect as expected. This keeps the behavior consistent with the expected outcome.

* update table component emptyText