Update fixes for MaintenancePlan
During add of new maintenance plan, we noticed that the
status selector does not invoke itself properly and throws error in console.
The useEffect was updating maintenance_status to be undefined and hence the error.
We fixed by applying a default status if the maintenance status is not found.

Update tech debt

The build was failing due to some lint issues. Resolved that as well.