Added storybook for LandingPage
Added storybook for landing page. We found out that the Authentication
components folder houses more folders, so in order to support that we just
decided to follow the same pattern in storybook as well. Seperated HeroContent
into its own mini component with text and button layouts and updated StyledAppBar
to suport the new redesign.