Add storybook for layout page (#301)
* Add storybook for layout page

* Added storybook for LandingPage

Added storybook for landing page. We found out that the Authentication
components folder houses more folders, so in order to support that we just
decided to follow the same pattern in storybook as well. Seperated HeroContent
into its own mini component with text and button layouts and updated StyledAppBar
to suport the new redesign.

* tech debt fixes

* Update Home section for storybook

Added redesign on the home section to make items more
visible and prominent in the screen. Added card components to
display all of the content and seperate it out into mini smaller components.
This gives us a clear abstraction and we were able to abstract items into
its own components easily.

Added navigation to the Asset Breakdown section so that users can easily click
and navigate to the desired location. The graph was also seperated out since it needed its own
component and keeping it like that would seamlessly transition itself into the
desired outcome in xs and sm mode as well.

Removed unused files esp the saga and slice as we did not find a use for it. Added some minor changes
to the RowHeader component and also updated paths for router. Included the ability to use no words if
stringToEdit was not passed in the utils function as well.

Added storybook for overview content files and used only the required args since this is tailored
directly to meet with the needs of the summary page