293 - Updated Asset list component structure
Updated asset list to contain some basic file structure changes initially.
Created new components to house AssetContent and created new directory to house
AssetListHeader files. Also updated AssetListHeader and AssetListContent files.
During this we also created mini components to handle asset movement. Created header and content
platform for basic drawer contents for viewing a selected asset. Added stories for AssetListContent
and AssetListHeader files. Renamed files to match the pattern we have been following so far.

Added changes to asset list

Renamed inventory list to asset list and there was just a lot of changes here
so decided to break down the changes slowly. Attempted to follow the pattern of naming
conventions that we have used so far, editing inventory is now known as SelectedAsset because
it is an asset and technically we are viewing it before we perform edits on it.

There is a lot of form components here at play, so we have not built any storybook for these sections
yet. We categorized assests to its own component and selected asset is its own component. By our given workflow
I think it would be best to seperate the saga and slice as well since they are two technically different things
and we also have to be mindful of not populating too much information on a single slice, which questions the
maintenability of the said architecture structure.

Fix changes to image picker

The image picker component in the Asset details page was not rendering correctly. We added change to support that
and fixed the issue. Added some cross functionality to render image during viewing assets at grid component level
and also viewing assets at their own SelectedAsset page. Basically, we just followed the pattern set for MaintenancePlan
since that seems like it was the most viable.

Fix tech debt and cleanup

A lot of files and folders have been moved around for easy cleanup and debugging purposes. Although the primary objective
of the ticket was to create storybook, due to so much traffic across the files, we will create storybook in a subsequent
pr and will link these two together.

Closes #293