Added changes to asset list
Renamed inventory list to asset list and there was just a lot of changes here
so decided to break down the changes slowly. Attempted to follow the pattern of naming
conventions that we have used so far, editing inventory is now known as SelectedAsset because
it is an asset and technically we are viewing it before we perform edits on it.

There is a lot of form components here at play, so we have not built any storybook for these sections
yet. We categorized assests to its own component and selected asset is its own component. By our given workflow
I think it would be best to seperate the saga and slice as well since they are two technically different things
and we also have to be mindful of not populating too much information on a single slice, which questions the
maintenability of the said architecture structure.