292 - Fix tech debt issues across the app
Fixed issue with categories and maintenance plan when sharing an image
there was api failure. That was due to misconfiguration on ui code that
didn't support required values.

Fixed issue with category list image page where the image is displayed
if the selected category has an image. This also fixed the issue of having
multiple same images being displayed in other categories.

Added material react v3 table for reports page to display reports and
cleaned up the checkbox actions. We should slowly migrate other table
components toward mrt v3 but awaiting mockups for that.

Added overflow to the existing table component so items do not overflow
the set height of 20rem. User can configure this height to be anything
they would like.

The drawer now shows favourite icons all the time. After the improvements
made on the drawer, the api callback was missing. Added this and it fixed
the issue.

Closes #292