Revamp Add bulk asset options
Added new layout for adding bulk asset upload

Added new icon to support upload of CSV. Added a different
layout to view the bulk upload. We also added the ability to view
file detials before the user can submit the file. Furthermore, if the
user decides that he / she does not want to upload the file, they can
do such by pressing the cancel button. Added relative position and
overflow on the card to align the cancel button. Created storybook
components for all the new files.

Created utility function to format date and time and also created utility
function to format the bytes of the file size in human readable format
so that the user is aware of the size of the file that they are trying
to parse and upload.