292 Added storybook for category page (#304)
* 292 Added storybook for category page

Since a lot of components were reusable, we decided that categories
and maintenance plans share the same components so they can be re-used
as much as possible. Refactored the category detail pages and maintenance
details pages.

Added section card for all card item headers

In categories and in maintenance plans we want to re-use the
section card component for the header and possibly for the content.
Working with the header atm.

Added section content and its relative stories.

Updated categories and maintenance plan to re-use the section card instead of having duplicate components

Refactor common components

Refactored a lot of common components between Categories
and Maintenance plan since they were quite similar
and we could just re-use them.

Fixed tech debt

- fixed some issues with the code

* fixed tech debt

* Add mrt v3 table

- Added mrt v3 table for DataTable.jsx file.

Added mrt v3 table and used it in reports component. Made everything
flush and works well with the existing reports page.

* 292 - Fix tech debt issues across the app

Fixed issue with categories and maintenance plan when sharing an image
there was api failure. That was due to misconfiguration on ui code that
didn't support required values.

Fixed issue with category list image page where the image is displayed
if the selected category has an image. This also fixed the issue of having
multiple same images being displayed in other categories.

Added material react v3 table for reports page to display reports and
cleaned up the checkbox actions. We should slowly migrate other table
components toward mrt v3 but awaiting mockups for that.

Added overflow to the existing table component so items do not overflow
the set height of 20rem. User can configure this height to be anything
they would like.

The drawer now shows favourite icons all the time. After the improvements
made on the drawer, the api callback was missing. Added this and it fixed
the issue.

Closes #292

* Tech debt cleanup

Add cleanup where the profile was not updating the correct time
when the appearance settings were changed. Updated the db layer
to handle the update of timestamp with time zone for the user modal.

Added cleanup where sharing 1 category with a third party was not allowing
the user to see that the item was being shared. rather the user had to
refresh the page. Fixed this issue so that the user does not have to
refresh the page anymore.

* Revamp Add bulk asset options

Added new layout for adding bulk asset upload

Added new icon to support upload of CSV. Added a different
layout to view the bulk upload. We also added the ability to view
file detials before the user can submit the file. Furthermore, if the
user decides that he / she does not want to upload the file, they can
do such by pressing the cancel button. Added relative position and
overflow on the card to align the cancel button. Created storybook
components for all the new files.

Created utility function to format date and time and also created utility
function to format the bytes of the file size in human readable format
so that the user is aware of the size of the file that they are trying
to parse and upload.

* Add fixes to images in Grid Component

* Tech debt changes

Found a couple of issues with the app when the user is new to the
system and does not have data. Fixes UI logic for the home page for this.