306 Update status list
Refine status list

Since we were refining our status list to only only
the provided status lists, we added some changes into the db section
so that the users are able to use the status list that was created
by the system. This allows the user to view the list of status and assign
categories or maintenance plans to it.

By doing such, we refrain the user from creating new statuses and let them
only use the existing ones to assign to the category or maintenance plan.
Although adding custom status sounds good and possibly a future use case, we
should refrain from that now since it just bloats the complexity and currently
the provided status list should be good enough.