Refactor update category between saga
Previously, we had moved update category to the
category item status and due to that, there were a couple of breaking
changes and bugs that prevented smooth removal and updates of our
selected category. Since there is no way to update the category from the
selected category page, we just added the saga and slice functions back to
the category page.