Allows category list to accurately display
Without this, when the user has one category, and attempts to remove
or update the selected category, the sorted data state will not be
updated since the function didn't get updated if there was no categories.
Instead we added code where if there was no categories, we wanted
sortedData state to accurately mimic the actual categories state from
the redux.