Update redux actions for maintenance plans
Updated redux actions for maintenance plans as well, since they
follow the same pattern like category as well.

Applied fix for bug where when the user has one maintenance plan,
removing it was not clearing it up from the ui and that was because the
sortedData state was not being updated when the length of maintenance plans
was zero.