Support seed of status in db
Status is list of enums used to filter and manage
categories and maintenance plans. We did not make this dynamic as the
overhead would be overkill for the feature. We have a list of statuses
that the user can use. If in the future, they require some custom status
then we will add additional tables to support adding / removing new
statuses at that time.

Updated prod docker files to support display of images from minio
containers and add seed data for prod instances as well. Update env vars
for production instance and fix prod.Dockerfile for frontend build as
it was in the incorrect context and was using Alpine node 18. Bumped it
up to node 20.

Closes #306