306 update status (#307)
* 306 Update status list

Refine status list

Since we were refining our status list to only only
the provided status lists, we added some changes into the db section
so that the users are able to use the status list that was created
by the system. This allows the user to view the list of status and assign
categories or maintenance plans to it.

By doing such, we refrain the user from creating new statuses and let them
only use the existing ones to assign to the category or maintenance plan.
Although adding custom status sounds good and possibly a future use case, we
should refrain from that now since it just bloats the complexity and currently
the provided status list should be good enough.

* Refactor update category between saga

Previously, we had moved update category to the
category item status and due to that, there were a couple of breaking
changes and bugs that prevented smooth removal and updates of our
selected category. Since there is no way to update the category from the
selected category page, we just added the saga and slice functions back to
the category page.

* Allows category list to accurately display

Without this, when the user has one category, and attempts to remove
or update the selected category, the sorted data state will not be
updated since the function didn't get updated if there was no categories.
Instead we added code where if there was no categories, we wanted
sortedData state to accurately mimic the actual categories state from
the redux.

* Update redux actions for maintenance plans

Updated redux actions for maintenance plans as well, since they
follow the same pattern like category as well.

Applied fix for bug where when the user has one maintenance plan,
removing it was not clearing it up from the ui and that was because the
sortedData state was not being updated when the length of maintenance plans
was zero.

* Update tech debt

Updated some tech debt issue where if the user types into the system
the data was not showing up in the ui.

* Update form component to display error

Updated Add form components to display error for Categories and
MaintenancePlans. Add required param to the TextFieldWithLabel
component and it will show a * at the end when the field is required

* Added tech debt changes for notes api

Added support for creating and updating notes for the first
time.

* Support seed of status in db

Status is list of enums used to filter and manage
categories and maintenance plans. We did not make this dynamic as the
overhead would be overkill for the feature. We have a list of statuses
that the user can use. If in the future, they require some custom status
then we will add additional tables to support adding / removing new
statuses at that time.

Updated prod docker files to support display of images from minio
containers and add seed data for prod instances as well. Update env vars
for production instance and fix prod.Dockerfile for frontend build as
it was in the incorrect context and was using Alpine node 18. Bumped it
up to node 20.

Closes #306

* add seed.data during test db

* update github runner with seed data

* Updates the OverviewHeader component

Updates the OverviewHeader component to display
list of categories that the items are against, instead of the list of
items. Since the list of items can be large, we want to display the list
of categories instead.

Follow similar steps for maintenance plans.

* rename wording for assets